correction of octal function

+++ format_functions_2.c
@@ -54,7 +54,7 @@ char *func_binary(va_list ap)
 char *func_octal(va_list ap)
 {
 	int nbytes, i;
-	long n, k;
+	unsigned int n, k;
 	char *ptr, *res;
 
 	k = va_arg(ap, long);
